+++ library/src/bundles/datastar-aliased.ts
@@ -12,23 +12,27 @@ import { RemoveFragments } from '../plugins/official/backend/watchers/removeFrag
 import { RemoveSignals } from '../plugins/official/backend/watchers/removeSignals'
 import { Clipboard } from '../plugins/official/browser/actions/clipboard'
 import { CustomValidity } from '../plugins/official/browser/attributes/customValidity'
-import { Intersects } from '../plugins/official/browser/attributes/intersects'
+import { OnIntersect } from '../plugins/official/browser/attributes/onIntersect'
+import { OnInterval } from '../plugins/official/browser/attributes/onInterval'
+import { OnLoad } from '../plugins/official/browser/attributes/onLoad'
+import { OnRaf } from '../plugins/official/browser/attributes/onRaf'
+import { OnSignalChange } from '../plugins/official/browser/attributes/onSignalChange'
 import { Persist } from '../plugins/official/browser/attributes/persist'
 import { ReplaceUrl } from '../plugins/official/browser/attributes/replaceUrl'
 import { ScrollIntoView } from '../plugins/official/browser/attributes/scrollIntoView'
-import { Show } from '../plugins/official/browser/attributes/show'
 import { ViewTransition } from '../plugins/official/browser/attributes/viewTransition'
 import { Attr } from '../plugins/official/dom/attributes/attr'
 import { Bind } from '../plugins/official/dom/attributes/bind'
 import { Class } from '../plugins/official/dom/attributes/class'
 import { On } from '../plugins/official/dom/attributes/on'
 import { Ref } from '../plugins/official/dom/attributes/ref'
+import { Show } from '../plugins/official/dom/attributes/show'
 import { Text } from '../plugins/official/dom/attributes/text'
 import { Fit } from '../plugins/official/logic/actions/fit'
 import { SetAll } from '../plugins/official/logic/actions/setAll'
 import { ToggleAll } from '../plugins/official/logic/actions/toggleAll'
 
-setAlias('ds')
+setAlias('star')
 
 load(
   // DOM
@@ -54,7 +58,11 @@ load(
   // Browser
   Clipboard,
   CustomValidity,
-  Intersects,
+  OnIntersect,
+  OnInterval,
+  OnLoad,
+  OnRaf,
+  OnSignalChange,
   Persist,
   ReplaceUrl,
   ScrollIntoView,