+++ library/src/plugins/official/backend/actions/sse.ts
@@ -2,29 +2,18 @@
 // Slug: Use a GET request to fetch data from a server using Server-Sent Events matching the Datastar SDK interface
 // Description: Remember, SSE is just a regular SSE request but with the ability to send 0-inf messages to the client.
 
-import { DATASTAR, DATASTAR_REQUEST, DefaultSseRetryDurationMs } from '../../../../engine/consts'
+import {
+  DATASTAR,
+  DATASTAR_REQUEST,
+  DefaultSseRetryDurationMs,
+} from '../../../../engine/consts'
 import { runtimeErr } from '../../../../engine/errors'
 import type { RuntimeContext } from '../../../../engine/types'
 import {
   type FetchEventSourceInit,
   fetchEventSource,
 } from '../../../../vendored/fetch-event-source'
-import {
-  DATASTAR_SSE_EVENT,
-  type DatastarSSEEvent,
-  ERROR,
-  FINISHED,
-  RETRYING,
-  STARTED,
-} from '../shared'
-
-function dispatchSSE(type: string, argsRaw: Record<string, string>) {
-  document.dispatchEvent(
-    new CustomEvent<DatastarSSEEvent>(DATASTAR_SSE_EVENT, {
-      detail: { type, argsRaw },
-    }),
-  )
-}
+import { dispatchSSE, ERROR, FINISHED, RETRYING, STARTED } from '../shared'
 
 const isWrongContent = (err: any) => `${err}`.includes('text/event-stream')
 
@@ -87,7 +76,7 @@ export const sse = async (
   const action = method.toLowerCase()
   let cleanupFn = (): void => {}
   try {
-    dispatchSSE(STARTED, { elId })
+    dispatchSSE(el, STARTED, { elId })
     if (!url?.length) {
       throw runtimeErr('SseNoUrlProvided', ctx, { action })
     }
@@ -113,7 +102,7 @@ export const sse = async (
       onopen: async (response: Response) => {
         if (response.status >= 400) {
           const status = response.status.toString()
-          dispatchSSE(ERROR, { status })
+          dispatchSSE(el, ERROR, { status })
         }
       },
       onmessage: (evt) => {
@@ -142,7 +131,7 @@ export const sse = async (
         }
 
         // if you aren't seeing your event you can debug by using this line in the console
-        dispatchSSE(type, argsRaw)
+        dispatchSSE(el, type, argsRaw)
       },
       onerror: (error) => {
         if (isWrongContent(error)) {
@@ -152,7 +141,7 @@ export const sse = async (
         // do nothing and it will retry
         if (error) {
           console.error(error.message)
-          dispatchSSE(RETRYING, { message: error.message })
+          dispatchSSE(el, RETRYING, { message: error.message })
         }
       },
     }
@@ -204,7 +193,7 @@ export const sse = async (
     urlInstance.search = queryParams.toString()
 
     try {
-      await fetchEventSource(ctx, urlInstance.toString(), req)
+      await fetchEventSource(el, urlInstance.toString(), req)
     } catch (error) {
       if (!isWrongContent(error)) {
         throw runtimeErr('SseFetchFailed', ctx, { method, url, error })
@@ -215,7 +204,7 @@ export const sse = async (
       // set the content-type to text/event-stream
     }
   } finally {
-    dispatchSSE(FINISHED, { elId })
+    dispatchSSE(el, FINISHED, { elId })
     cleanupFn()
   }
 }