+++ library/src/plugins/official/backend/watchers/mergeFragments.ts
@@ -1,10 +1,10 @@
-// Icon: material-symbols:cloud-download
+// Authors: Delaney Gillilan
+// Icon: material-symbols:sim-card-download-outline
 // Slug: Merge fragments into the DOM using a Server-Sent Event
 // Description: Remember, SSE is just a regular SSE request but with the ability to send 0-inf messages to the client.
 
 import {
   DefaultFragmentMergeMode,
-  DefaultFragmentsSettleDurationMs,
   DefaultFragmentsUseViewTransitions,
   EventTypes,
   FragmentMergeModes,
@@ -23,11 +23,7 @@ import {
   supportsViewTransitions,
 } from '../../../../utils/view-transtions'
 import { Idiomorph } from '../../../../vendored/idiomorph.esm'
-import {
-  SETTLING_CLASS,
-  SWAPPING_CLASS,
-  datastarSSEEventWatcher,
-} from '../shared'
+import { datastarSSEEventWatcher } from '../shared'
 
 export const MergeFragments: WatcherPlugin = {
   type: PluginType.Watcher,
@@ -40,12 +36,9 @@ export const MergeFragments: WatcherPlugin = {
         fragments: fragmentsRaw = '<div></div>',
         selector = '',
         mergeMode = DefaultFragmentMergeMode,
-        settleDuration:
-          settleDurationRaw = `${DefaultFragmentsSettleDurationMs}`,
         useViewTransition:
           useViewTransitionRaw = `${DefaultFragmentsUseViewTransitions}`,
       }) => {
-        const settleDuration = Number.parseInt(settleDurationRaw)
         const useViewTransition = isBoolString(useViewTransitionRaw)
 
         fragmentContainer.innerHTML = fragmentsRaw.trim()
@@ -63,10 +56,10 @@ export const MergeFragments: WatcherPlugin = {
 
           if (useViewTransition && supportsViewTransitions) {
             docWithViewTransitionAPI.startViewTransition(() =>
-              applyToTargets(ctx, mergeMode, settleDuration, fragment, targets),
+              applyToTargets(ctx, mergeMode, fragment, targets),
             )
           } else {
-            applyToTargets(ctx, mergeMode, settleDuration, fragment, targets)
+            applyToTargets(ctx, mergeMode, fragment, targets)
           }
         }
       },
@@ -77,18 +70,19 @@ export const MergeFragments: WatcherPlugin = {
 function applyToTargets(
   ctx: InitContext,
   mergeMode: string,
-  settleDuration: number,
   fragment: Element,
   capturedTargets: Element[],
 ) {
-  for (const initialTarget of capturedTargets) {
-    initialTarget.classList.add(SWAPPING_CLASS)
-    const originalHTML = initialTarget.outerHTML
-    const modifiedTarget = initialTarget
+  for (const target of capturedTargets) {
+    // Mark the target as a fragment merge target to force plugins to clean up and reapply
+    (target as HTMLElement).dataset.fragmentMergeTarget = 'true'
+
+    // Clone the fragment to merge to avoid modifying the original and force browsers to merge the fragment into the DOM
+    const fragmentToMerge = fragment.cloneNode(true) as HTMLorSVGElement
+
     switch (mergeMode) {
       case FragmentMergeModes.Morph: {
-        const fragmentWithIDs = fragment.cloneNode(true) as HTMLorSVGElement
-        walkDOM(fragmentWithIDs, (el) => {
+        walkDOM(fragmentToMerge, (el) => {
           if (!el.id?.length && Object.keys(el.dataset).length) {
             el.id = elUniqId(el)
           }
@@ -105,59 +99,42 @@ function applyToTargets(
           }
         })
 
-        Idiomorph.morph(modifiedTarget, fragmentWithIDs)
+        Idiomorph.morph(target, fragmentToMerge)
         break
       }
       case FragmentMergeModes.Inner:
         // Replace the contents of the target element with the outer HTML of the response
-        modifiedTarget.innerHTML = fragment.outerHTML
+        target.innerHTML = fragmentToMerge.outerHTML
         break
       case FragmentMergeModes.Outer:
         // Replace the entire target element with the response
-        modifiedTarget.replaceWith(fragment)
+        target.replaceWith(fragmentToMerge)
         break
       case FragmentMergeModes.Prepend:
         // Insert the response before the first child of the target element
-        modifiedTarget.prepend(fragment)
+        target.prepend(fragmentToMerge)
         break
       case FragmentMergeModes.Append:
         // Insert the response after the last child of the target element
-        modifiedTarget.append(fragment)
+        target.append(fragmentToMerge)
         break
       case FragmentMergeModes.Before:
         // Insert the response before the target element
-        modifiedTarget.before(fragment)
+        target.before(fragmentToMerge)
         break
       case FragmentMergeModes.After:
         // Insert the response after the target element
-        modifiedTarget.after(fragment)
+        target.after(fragmentToMerge)
         break
       case FragmentMergeModes.UpsertAttributes:
         // Upsert the attributes of the target element
-        for (const attrName of fragment.getAttributeNames()) {
-          const value = fragment.getAttribute(attrName)!
-          modifiedTarget.setAttribute(attrName, value)
+        for (const attrName of fragmentToMerge.getAttributeNames()) {
+          const value = fragmentToMerge.getAttribute(attrName)!
+          target.setAttribute(attrName, value)
         }
         break
       default:
         throw initErr('InvalidMergeMode', ctx, { mergeMode })
-    }
-
-    const cl = modifiedTarget.classList
-    cl?.add(SWAPPING_CLASS)
-
-    setTimeout(() => {
-      initialTarget.classList.remove(SWAPPING_CLASS)
-      cl?.remove(SWAPPING_CLASS)
-    }, settleDuration)
-
-    const revisedHTML = modifiedTarget.outerHTML
-
-    if (cl && originalHTML !== revisedHTML) {
-      cl.add(SETTLING_CLASS)
-      setTimeout(() => {
-        cl.remove(SETTLING_CLASS)
-      }, settleDuration)
     }
   }
 }