+++ site/smoketests/signals_change_test.go
@@ -1,9 +1,0 @@
-package smoketests
-
-import (
-	"testing"
-)
-
-func TestUnitSignalsChange(t *testing.T) {
-	setupPageTestOnClick(t, "tests/signals_change")
-}